+++ platforms/forge-1.18.2/src/main/java/org/anti_ad/mc/ipnext/ingame/Inventory.kt
@@ -45,7 +45,8 @@ fun vCursorStack() = Vanilla.playerInventory().player.containerMenu.carried?.`(i
 fun vPlayerSlotOf(slot: Slot,
                   screen: Screen?): Slot { // creative slot to survival slot
     if (screen !is CreativeInventoryScreen) return slot
-    if (slot.`(inventory)` !is PlayerInventory) return slot
+    val inventory = slot.`(inventoryOrNull)`
+    if (inventory != null && inventory !is PlayerInventory) return slot
     val id = slot.`(id)`
     val invSlot = slot.`(invSlot)`
     Log.trace("id: $id, invSlot: $invSlot")